Update service.py
The user name is stored in `username` rather than `user_name`.

Side note: the JWT is already parsed before the service is run. Hence, I think it would be better to rely on the `user` field from the `auth_env` section that contains the username, whatever the name of the key storing the user name.